refactor btc-backend, fix response parsing
